requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-multipart
PyJWT
mysql-connector-python